claude-test/SwarmCare/requirements.txt
# Core Dependencies
crewai>=0.70.0
crewai-tools>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Azure Dependencies
openai>=1.54.0
azure-ai-contentsafety>=1.0.0
azure-identity>=1.18.0
azure-core>=1.30.0

# Monitoring and Logging
opencensus-ext-azure>=1.1.13
opencensus>=0.11.4

# Medical NLP and Validation
spacy>=3.7.0
transformers>=4.44.0

# Retry and Resilience
tenacity>=9.0.0
backoff>=2.2.1

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# Utilities
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.10.0
pyyaml>=6.0.0
